requirements.txt
imutils==0.5.4
numpy==1.26.4
opencv-python==4.9.0.80
python-dotenv==1.0.1
